refactor:informationコントローラーをadminの直下に修正。
